feat: add sound when breakpoint set or removed
By default IntelliJ gives no feedback when breakpoints are added or removed so now we play a sound for added and another for removing. We're using a deprecated method as for the life of us we can't figure out the right types to listen to any breakpoint event.

In theory we should be able to use plugin.xml to add the listener rather than appStarted but that doesn't work (nothing happens when breakpoints are toggled)

Getting the current project from appStarted is hard work too, technically there might not be one but we cross our fingers at the moment