feat: customise sounds
You can now use the settings dialog to specify a path to a folder which contains wav files to override the built in sounds. The possible file names are defined in the enum Sound. You can also toggle warning sounds on/off with the dialog or a shortcut

We use Files.exists so if the current sound isn't in the custom folder we can use the default which stops people needing to replace all the sounds at once

The settings dialog has been pretty much copied from the intellij example code https://github.com/JetBrains/intellij-sdk-code-samples/tree/main/settings