fix: beep and say issues that expand the whole file
When you have certain issues with files like yaml (e.g a GitHub action has a step instead of a steps key or is missing jobs) the warning expands from the start of the issue to the end of the file which means the previous "is issue on current line" logic didn't work as the line end offset was less than the highlight end offset. In this case we still want people to know there's a problem, so now we check if it's on this current line or within the text range to ensure multi line issues are picked up.
We kept the "is on current line" logic so warning beeps can be played when you enter a line with an issue but not the caret isn't precisely on the issue

Got a new Position class to reduce a bit of duplicated logic so now fixes to that logic will fix both the say command and beeps. Also if the errorStripeTooltip is null we might as well bail from the highlighter loop as early as possible for efficiency