chore(refactor): Move all queue references into sayText
Jetbrains warns us

Never acquire service instances prematurely or store them in fields for later use. Instead, always obtain these service instances directly and only at the location where they're needed. Failing to do so can lead to unexpected exceptions and severe consequences for your plugin's functionality.

By passing the queue from the DebugListener through to the various callbacks we were ignoring this advice, now we only fetch the queue at the point we need it and have wrapped it all up inside the sayText method, with the original sayText being renamed to the more accurate createSpeech

This is a  bit convoluted as you jump back and forth between Speech and SpeechQueue to follow the code but it'll do for now and centralises the queue running logic