chore(test): Added a test!
Testing the most basic functionality, toggling a setting by triggering an action. We've yet to figure out how to simulate keyboard shortcuts so this is a good start.

We've mocked the audible queue to be able to assert something is said without actually speaking. We'll do the same for sound too