chore(test): Some decent test coverage
Not all scenarios covered but enough to save some time!

BaseTestCase.kt now resets the mocks as they don't get recreated for each test.

We'd like to test the warning beeps but that requires setting up a mock jdk which we haven't quite got working yet so we'll settle for errors