fix(tests): reset breakpoints after test
Otherwise they bleed into other tests which causes things like BreakpointListenerTest to fail.

Also the test that checked if sound was disabled on a breakpoint was passing because we weren't setting a breakpoint as we learnt in c24a9911b11c2cbf82e4f149bde530c40ff4c853 oops. Now we are setting a breakpoint so it should fail if the setting code is changed