feat: use voiceover on Mac by default
A lot of people use VoiceOver with Mac so it makes sense that IntelliSay would use that when available. The simplest way is to execute AppleScript via the command line. This means a user has to enable VoiceOver control from AppleScript. This is potentially a security issue so we will look into using the accessibility API instead in future updates which we think will work around this. But as a first draft it is a good start!

We used the detection method so users don't have to set a setting in IntelliSay but we will make this available in the future