feat: announce dumb mode status
When intellij is first opened it will often enter dumb mode a few times which means most features are not available and this can be quite confusing if you are not aware that this has happened. Now we announce when you enter and exit dumb mode.

This could potentially be annoying so there is a setting to disable it

Testing the exit announcement is more complicated as it runs asynchronously so it happens after the test has finished.