feat: user configurable voiceover
Now via the UI you can toggle voiceover on or off.

The initial check is smarter and if you want voiceover (default on Mac) and don't have AppleScript configured we will use the built in say command to tell the user that they need to enable it, then fallback to that.