feat: announce current method and class
Adding a python plugin dependency as you need to search for a PyFunction rather than a PSIMethod (PSIMethod works for Java only it seems)
It needs to be optional else the unit tests break as no services are created (and therefore no speech/sound). Honestly don't know why
but any jetbrains plugin caused this to happen.

It being optional should be fine as the language is not identified by IntelliJ and comes through as the default language (probably TextMate)
when on a python file. We do a crude check for the file extension to indicate to the user they need to enable the python plugin