feat: announce current method and class
Adding a python plugin dependency as you need to search for a PyFunction rather than a PSIMethod (PSIMethod works for Java only it seems)
It needs to be optional else the unit tests break as no services are created (and therefore no speech/sound). Honestly don't know why
but any jetbrains plugin caused this to happen.

It being optional should be fine if you don't use the python announce context