feat: set python interpreter
The current method to set an interpreter in IntelliJ is painful as it involves up to 3 dialogues which are all difficult to navigate and understand with a screen reader.

This shortcut simplifies the process to one dialogue with a text input with the path to the location of the python interpreter either as a full path to the python executable or a folder containing a bin directory with the python executable.

As it only applies to python it is in the python XML so if you don't have the python plug-in, it won't be loaded

We default to venv as it's a common python interpreter manager

Code largely copied from IntelliJ source code[1]

[1]: https://github.com/JetBrains/intellij-community/blob/0ffa5e48a05b3f09226a92c3acdf02c9350ce771/plugins/evaluation-plugin/languages/python/src/com/intellij/cce/evaluation/SetupPythonInterpreterStep.kt#L18