feat: help dialog
You can now open help with command shift option h and choose an action to take.

We use a custom renderer to be able to handle rendering actions descriptively.

It's annoying for Mac users to have to mentally switch meta to command and alt to option so replace them in the string if we're on mac