refactor: make dialogue and sdk creation part of a service to be mockable
You can't open dialogs in unit tests in IntelliJ plugins so we've created a service to mock that in the test

Loading a full Python plugin is not possible either[1]. The recommendation is to copy a lot of code from the IntelliJ repo (starting at PyEnvTestCase) or implement your own version. Both are a lot of overhead and maintenance so we just mock the part that throws an 'Invalid JDK type: PythonSDK' and use Java as the type, as that's available.

To avoid a 'leaked sdk' error we need to override reset in the mocksdkhelper and remove anything we might have created

[1]: https://intellij-support.jetbrains.com/hc/en-us/community/posts/20913926057746/comments/21133012763282