tests: test setting interpreter action
The "existing sdk" test needs to remove the sdk it registers before calling the action, else we get a 'leaked sdk' error

As the code converts the local path to an absolute path and the tests project.basePath is a generated temp directory we can't match the exact output for speech reliably so we use a fuzzy match with a regex to match the generated temp dir.
Because the basePath is generated and is also empty we can't cover the "local interpreter and valid" case. At least not yet, maybe we'll figure that out in the future

The python file doesn't need to be a valid python executable for the test, just to exist, hence empty file