fix: linux audio crashes when navigating quickly
Fixes #1

If you moved your cursor real fast, or toggled a breakpoint a lot the audio would crash as it was taking lines and not releasing them. It was also unreliable at selecting which audio output to use.

AudioCue[1] from philfrei allows us easier control over audio and some more reliable sound output.

Copied the source code in as I know some companies like to audit code before allowing it to be used and this makes that process easier. Auto conversion was performed, with some tweaks of arrays becoming arrayLists to avoid needed to make the items nullable and changing lots of code.

Volume will be adjustable in an upcoming commit, hard coded at full volume for now.

[1]: https://github.com/philfrei/AudioCue-maven/tree/main